+++ semilearn/datasets/cv_datasets/datasetbase.py
@@ -85,7 +85,11 @@ class BasicDataset(Dataset):
                 img = Image.fromarray(img)
             img_w = self.transform(img)
             if not self.is_ulb:
-                return {'idx_lb': idx, 'x_lb': img_w, 'y_lb': target} 
+                if self.alg == 'defixmatch' and self.strong_transform is not None:
+                    # NOTE Strong augmentation on the labelled for DeFixMatch
+                    return {'idx_lb': idx, 'x_lb': img_w, 'x_lb_s': self.strong_transform(img), 'y_lb': target} 
+                else:
+                    return {'idx_lb': idx, 'x_lb': img_w, 'y_lb': target}
             else:
                 if self.alg == 'fullysupervised' or self.alg == 'supervised':
                     return {'idx_ulb': idx}