+++ CHANGE_LOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 23/07/15/2023 Update
+
+* Added DeFixmatch algorithm.
+
 ## 23/01/2023 Update
 
 * Added imbalanced Semi-Supervised setting and 8 imbalanced algorithms.